requirements.txt
# Core Dependencies
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0
tensorboard>=2.13.0

# Data and utilities
scipy>=1.10.0
scikit-learn>=1.3.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
